Crear endpoint Clima de Ciudades
- Eliminar endpoint Clima de Ciudades por Coordenadas.
- Añadir endpoint Clima de Ciudades por ID de Ciudad.
- Añadir Base de Datos weather-service.sql.
- Crear Exception personalizada CityNotFoundException.

+++ weather-service/src/main/java/com/cibertec/client/WeatherClient.java
@@ -6,6 +6,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.math.BigDecimal;
+
 @FeignClient(name = "weatherApiClient", url = "${weather.api.base-url}")
 public interface WeatherClient {
 
@@ -14,8 +16,8 @@ public interface WeatherClient {
             produces = { "application/json" }
     )
     ResponseEntity<WeatherApiResponse> getWeatherData(
-            @RequestParam Double lat,
-            @RequestParam Double lon,
+            @RequestParam BigDecimal lat,
+            @RequestParam BigDecimal lon,
             @RequestParam String appid,
             @RequestParam String lang,
             @RequestParam String units
